% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreebar-package.R
\docType{package}
\name{ggtreebar-package}
\alias{ggtreebar}
\alias{ggtreebar-package}
\title{ggtreebar: Make Treemap Bar Charts with 'ggplot2'}
\description{
'ggplot2' geoms analogous to 'geom_col()' and 'geom_bar()' that allow for treemaps using 'treemapify' nested within each bar segment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/ggtreebar}
  \item Report bugs at \url{https://github.com/hrryt/ggtreebar/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
